+++ tests/core/transmission/trans_channel/tcp_direct/unittest/mock/trans_tcp_direct_message_test_mock.h
@@ -77,7 +77,7 @@ public:
     virtual int32_t TransTdcOnMsgReceived(const char *pkgName, int32_t pid,
         int32_t channelId, TransReceiveData *receiveData) = 0;
     virtual int32_t LnnGetRemoteNodeInfoById(const char *id, IdCategory type, NodeInfo *info) = 0;
-    virtual int32_t TransCheckServerAccessControl(uint32_t callingTokenId) = 0;
+    virtual int32_t TransCheckServerAccessControl(uint64_t callingTokenId) = 0;
     virtual int32_t TransTdcOnChannelClosed(const char *pkgName, int32_t pid, int32_t channelId) = 0;
 };
 
@@ -127,7 +127,7 @@ public:
     MOCK_METHOD4(TransTdcOnMsgReceived, int32_t (const char *pkgName, int32_t pid,
         int32_t channelId, TransReceiveData *receiveData));
     MOCK_METHOD3(LnnGetRemoteNodeInfoById, int32_t (const char *id, IdCategory type, NodeInfo *info));
-    MOCK_METHOD1(TransCheckServerAccessControl, int32_t (uint32_t callingTokenId));
+    MOCK_METHOD1(TransCheckServerAccessControl, int64_t (uint64_t callingTokenId));
     MOCK_METHOD3(TransTdcOnChannelClosed, int32_t (const char *pkgName, int32_t pid, int32_t channelId));
 };
 } // namespace OHOS
